VOL-4499: Update 'device_flow' struct to have the flow_type. This will enable the voltha openolt adapter to just send the voltha-flow-id and the agent can retrieve the flow_type automatically from the device_flow struct.
VOL-4520: Allow configurability of MAC Device Activation delay time.
With BAL3.10.2.2 it is observed the Edgecore OLTs sometimes need more
time to do the Maple/Aspen download procedure as part of MAC activation.
The current 200ms delay has recomendded to increase to 2s as a
workaround by Edgecore (ticket #19146). This patch makes this delay
configurable on a per platform basis.

Version bumped to 3.6.3

Change-Id: I3bac3af423aa6374c10302519cd21a481d5a4a86